project_root/requirements.txt
pandas
sqlalchemy
psycopg2
python-dotenv
sentence-transformers
scikit-learn
openai
requests
